Add NVMe and RTL8139 drivers, enhance syscall handling
- Added NVMe and RTL8139 drivers to the project, including necessary dependencies in Cargo files.
- Introduced new syscalls for handle management: `SYS_HANDLE_GRANT` and `SYS_HANDLE_REVOKE`.
- Enhanced the syscall dispatcher to support the new handle management syscalls.
- Updated memory management functions to include a new `allocate_dma_frame` for DMA operations.
- Improved NVMe driver implementation with corrected memory address handling and initialization in the kernel main function.

+++ workspace/drivers/storage/nvme/src/lib.rs
@@ -230,7 +230,7 @@ impl NvmeController {
             .ok_or(NvmeError::IoError)?;
 
         let phys = frame.start_address() as u64;
-        let virt = crate::memory::phys_to_virt(phys as usize) as *mut u8;
+        let virt = crate::memory::phys_to_virt(phys) as *mut u8;
 
         unsafe {
             ptr::write_bytes(virt, 0, NVME_PAGE_SIZE);
@@ -311,7 +311,7 @@ impl NvmeController {
             .ok_or(NvmeError::IoError)?;
 
         let phys = frame.start_address() as u64;
-        let virt = crate::memory::phys_to_virt(phys as usize) as *mut u8;
+        let virt = crate::memory::phys_to_virt(phys) as *mut u8;
 
         let blocks = (buffer.len() / 512) as u32;
 
@@ -350,7 +350,7 @@ impl NvmeController {
             .ok_or(NvmeError::IoError)?;
 
         let phys = frame.start_address() as u64;
-        let virt = crate::memory::phys_to_virt(phys as usize) as *mut u8;
+        let virt = crate::memory::phys_to_virt(phys) as *mut u8;
 
         unsafe {
             ptr::copy_nonoverlapping(buffer.as_ptr(), virt, buffer.len());
